requirements.txt
# AI & LLM
langchain==0.3.7
langchain-community==0.3.7
langgraph==0.2.50
google-genai==1.0.0
anthropic==0.39.0

# Web Search & Scraping
duckduckgo-search==6.3.7
requests==2.32.3
beautifulsoup4==4.12.3
httpx==0.27.2

# Vector Database
faiss-cpu==1.9.0
sentence-transformers==3.3.1

# Data Analysis
pandas==2.2.3
numpy==2.1.3
matplotlib==3.9.3
seaborn==0.13.2
plotly==5.24.1

# API Backend
fastapi==0.115.5
uvicorn==0.32.1
python-multipart==0.0.17
pydantic==2.10.1

# Document Generation
reportlab==4.2.5
fpdf2==2.8.1
python-docx==1.1.2

# Utilities
python-dotenv==1.0.1
schedule==1.2.2
rich==13.9.4
loguru==0.7.3
tenacity==9.0.0
aiofiles==24.1.0
